+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix course run synchronization when triggered on enrollment
+
 ## [5.12.0] - 2022-01-17
 
 ### Added